% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalizedCovariance.R
\name{penalizedCovariance}
\alias{penalizedCovariance}
\title{Compute the coefficients for the covariance kernel in a product bspline basis}
\usage{
penalizedCovariance(
  indexes,
  values,
  lower,
  upper,
  nbasis,
  order = 4,
  splinePenalty,
  weights = NULL,
  lambda = 10,
  use_cv = FALSE,
  mixedOnly = FALSE,
  lower_optim = 10^0.25,
  upper_optim = 10^6
)
}
\arguments{
\item{indexes}{List of indexes of the quantity observed as a random function}

\item{values}{List of values of the quantity observed as a random function at indexes}

\item{lower}{Lower interval boundary for the splines}

\item{upper}{Upper interval boundary for the splines}

\item{nbasis}{Number of basis functions}

\item{order}{The order of the bsplines (default 4)}

\item{splinePenalty}{Penalty Matrix for one bspline basis}

\item{weights}{Vector of weights for the function observations (default equal weights)}

\item{lambda}{Penalty parameter (default 10)}

\item{use_cv}{Whether to use generalized cross validation to determine the penalty parameter (default FALSE)}

\item{mixedOnly}{Whether to exclude estimates for the second moments(i.e. only values[indexes[i]]^2 or only values[indexes[i]] * values[indexes[i]], i != j) for the estimation}

\item{lower_optim}{Parameter for optimization during gcv, (default = 10^0.25)}

\item{upper_optim}{Parameter for optimization during gcv, (upper_optim = 10^6)}
}
\value{
A matrix of coefficients for the product bspline basis
}
\description{
This function takes lists of indexes and values of observations of random functions
and estimates the marginal covariance kernel in a product bspline basis
}
