% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalizedCrossCovariance.R
\name{penalizedCrossCovariance}
\alias{penalizedCrossCovariance}
\title{Compute the coefficients for the cross-covariance kernel in a product bspline basis}
\usage{
penalizedCrossCovariance(
  indexes1,
  values1,
  indexes2,
  values2,
  intv1,
  intv2,
  nbasis,
  order = as.integer(c(4, 4)),
  weights = NULL,
  lambda = 10,
  use_cv = FALSE,
  lower_optim = 10^0.25,
  upper_optim = 10^6
)
}
\arguments{
\item{indexes1}{List of indexes of the first quantity observed as a random function}

\item{values1}{List of values of the first quantity observed as a random function}

\item{indexes2}{List of indexes of the second quantity observed as a random function}

\item{values2}{List of values of the second quantity observed as a random function}

\item{intv1}{A vector of length two containing the interval boundaries for the splines for first quantity}

\item{intv2}{A vector of length two containing the interval boundaries for the splines for second quantity}

\item{nbasis}{Number of basis functions (currently the same for both quantities)}

\item{order}{Integer vector of length two determining the order of the bsplines (default c(4,4))}

\item{weights}{Vector of weights for the function observations (default equal weights)}

\item{lambda}{Penalty parameter (default 10)}

\item{use_cv}{Whether to use generalized cross validation to determine the penalty parameter (default FALSE)}

\item{lower_optim}{Parameter for optimization during gcv, (default = 10^0.25)}

\item{upper_optim}{Parameter for optimization during gcv, (upper_optim = 10^6)}
}
\value{
A matrix of coefficients for the product bspline basis
}
\description{
This function takes lists of indexes and values of observations of random functions
and estimates the marginal cross-covariance kernel in a product bspline basis
}
